+++ temba/flows/migrations/0081_install_triggers.py
@@ -5,6 +5,7 @@ from __future__ import unicode_literals
 from django.db import migrations
 from temba.sql import InstallSQL
 
+
 class Migration(migrations.Migration):
 
     dependencies = [