+++ temba/perf_tests.py
@@ -1,4 +1,4 @@
-from __future__ import unicode_literals
+from __future__ import print_function, unicode_literals
 
 import json
 import random
@@ -62,9 +62,9 @@ class PerformanceTest(TembaTest):  # pragma: no cover
 
     @classmethod
     def tearDownClass(cls):
-        print "\n------------------ Segment Profiles ------------------"
+        print("\n------------------ Segment Profiles ------------------")
         for segment in cls.segments:
-            print six.text_type(segment)
+            print(six.text_type(segment))
 
     def _create_contacts(self, count, base_names):
         """